571: v 2.0: restructure db_types
