+++ app/db/types/Consumers.py
@@ -1,0 +1,10 @@
+from app.db.types.Packable import Packable
+
+
+class Consumers(Packable):
+    def __init__(self, dictionary=None):
+        super().__init__(dictionary)
+        dictionary = dictionary or {}
+        self.consumer_key = dictionary.get('consumer_key', '')
+        self.consumer_secret = dictionary.get('consumer_secret', '')
+        self.timestamp_and_nonce = dictionary.get('timestamp_and_nonce', [])
