+++ app/db/types/Packable.py
@@ -1,0 +1,9 @@
+class Packable:
+    def __init__(self, dictionary):
+        pass
+
+    def pack(self):
+        return dict(vars(self))
+
+    def __str__(self) -> str:
+        return f"{self.__class__.__name__}: {', '.join([f'{key}: {value}' for key, value in vars(self).items()])}"
